backend/requirements.txt
asgiref==3.8.1
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
fastapi==0.111.0
pydantic==2.7.4
sqlparse==0.5.0
uvicorn==0.30.1
